+++ README.md
@@ -247,7 +247,11 @@ API возвращает соответствующие HTTP статус-код
 ### Структура проекта
 ```
 todo-api/
-├── main.go          # Основной файл с сервером и обработчиками
+├── main.go          # Основной файл с точкой входа
+├── models.go        # Модели данных (Task, CreateTaskRequest, UpdateTaskRequest, ErrorResponse)
+├── services.go      # Интерфейс TaskServiceInterface и реализация TaskService
+├── handlers.go      # HTTP обработчики (TaskHandler)
+├── routes.go        # Настройка маршрутов и middleware
 ├── main_test.go     # Юнит-тесты
 ├── go.mod           # Модуль Go
 ├── go.sum           # Хеши зависимостей
@@ -256,10 +260,18 @@ todo-api/
 
 ### Компоненты
 
-1. **Task** - структура данных задачи
-2. **TaskService** - бизнес-логика для работы с задачами
-3. **TaskHandler** - HTTP обработчики
-4. **Middleware** - промежуточное ПО (логирование, CORS, восстановление)
+1. **Models** (`models.go`) - структуры данных для задач и запросов
+2. **Services** (`services.go`) - бизнес-логика с интерфейсом TaskServiceInterface
+3. **Handlers** (`handlers.go`) - HTTP обработчики для REST API
+4. **Routes** (`routes.go`) - настройка маршрутов и middleware
+5. **Main** (`main.go`) - точка входа и инициализация приложения
+
+### Принципы архитектуры
+
+- **Разделение ответственности** - каждый файл отвечает за свою область
+- **Интерфейсы** - TaskService возвращает интерфейс, что обеспечивает гибкость и тестируемость
+- **Dependency Injection** - зависимости передаются через конструкторы
+- **Чистая архитектура** - четкое разделение между слоями приложения
 
 ## 🔒 Безопасность
